+++ src/index.tsx
@@ -4,10 +4,15 @@ import App from './components/app/app';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import ErrorMessage from './components/error-message/error-message';
-import { checkAuthAction, fetchOfferAction } from './store/api-actions';
+import { checkAuthAction, fetchOfferAction, getUserData } from './store/api-actions';
+import { getToken } from './services/token';
 
 store.dispatch(fetchOfferAction());
 store.dispatch(checkAuthAction());
+
+if (getToken()) {
+  store.dispatch(getUserData());
+}
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
