+++ src/store/slices/auth-slice/auth-action.ts
@@ -1,0 +1,4 @@
+import { createAction } from '@reduxjs/toolkit';
+import { AuthorizationStatus } from '../../../constants';
+
+export const requireAuthorization = createAction<AuthorizationStatus>('user/requireAuthorization');
