+++ src/store/slices/error-slice/error-action.ts
@@ -1,0 +1,3 @@
+import { createAction } from '@reduxjs/toolkit';
+
+export const setError = createAction<string | null>('/setError');
