+++ src/store/slices/sorting-slice/sorting-action.ts
@@ -1,0 +1,4 @@
+import { createAction } from '@reduxjs/toolkit';
+import { SortTypes } from '../../../types/models';
+
+export const setSorting = createAction<SortTypes>('/sorting');
