+++ src/store/slices/town-slice/town-action.ts
@@ -1,0 +1,4 @@
+import { createAction } from '@reduxjs/toolkit';
+import { City } from '../../../types/models';
+
+export const changeTown = createAction<City>('/changeTown');
