+++ src/store/slices/town-slice/town-slice.tsx
@@ -1,6 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { CITIES } from '../../../constants';
-import { changeTown } from '../../action';
+import { changeTown } from './town-action';
+
 
 export const townsSlice = createSlice({
   name: 'towns',
